refactor(path): separate workspace validation from path normalization (#206) Split path/normalize.ts (660 lines) into: - path/normalize.ts (284 lines) - path normalization, tilde expansion, SSH key validation - path/workspace.ts (399 lines) - workspace/submodule path validation No circular dependencies. All APIs remain compatible via pathUtils.ts re-exports. All tests pass.
🖥️ IDE: [Cursor](https://cursor.sh)
🔌 Extension: [Claude Code](https://claude.ai/download)

Model-Raw: claude-opus-4-5-20251101
