+++ extensions/git-id-switcher/src/pathUtils.ts
@@ -4,14 +4,13 @@
  * This module re-exports functions from the path/ directory for backwards compatibility.
  * New code should import directly from the specific modules:
  * - './path/normalize' for path normalization
+ * - './path/workspace' for workspace/submodule validation
  * - './path/symlink' for symlink operations
  */
 
-// Re-export all public APIs from normalize module
+// Re-export path normalization APIs
 export {
   normalizeAndValidatePath,
-  validateWorkspacePath,
-  validateSubmodulePath,
   validateSshKeyPath,
   expandTilde,
 } from './path/normalize';
@@ -20,8 +19,16 @@ export {
 export type {
   NormalizePathOptions,
   NormalizedPathResult,
-  ValidateSubmodulePathOptions,
 } from './path/normalize';
+
+// Re-export workspace validation APIs
+export {
+  validateWorkspacePath,
+  validateSubmodulePath,
+} from './path/workspace';
+
+// Re-export types from workspace module
+export type { ValidateSubmodulePathOptions } from './path/workspace';
 
 // Re-export symlink functions
 export { containsSymlinks } from './path/symlink';
